Evolved API modeling

+++ src/main/java/com/brunoreolon/cinebaianosapi/api/model/vote/stats/VoteStatsResponse.java
@@ -13,6 +13,6 @@ import lombok.Setter;
 public class VoteStatsResponse {
 
     private VoteTypeSummaryResponse type;
-    private Integer totalVotes;
+    private Long totalVotes;
 
 }
